+++ src/components/Stations.tsx
@@ -3,6 +3,7 @@ import { Box, FormControl, InputLabel, MenuItem } from '@mui/material';
 import Select, { SelectChangeEvent } from '@mui/material/Select';
 import { fetchQueryInformation } from '../helpers/ApiCallHelper';
 import Station from './Station';
+import { StationDropdown } from './StationDropdown';
 
 const Stations: React.FC = () => {
 
@@ -35,73 +36,27 @@ const Stations: React.FC = () => {
             .finally(() => console.log(params.toString()));
     };
 
-    const stationList = [
-        { value: 'PAD', label: 'Paddington' },
-        { value: 'LST', label: 'Liverpool Street' },
-        { value: 'KGX', label: 'Kings Cross' },
-        { value: 'KTN', label: 'Kentish Town' },
-        { value: 'LBG', label: 'London Bridge' },
-    ] as Array<{value: string; label: string}>;
-
     const [departureStationChoice, setDepartureStationChoices] = useState('');
     const [arrivalStationChoice, setArrivalStationChoices] = useState('');
-
-    const handleChange = (event: SelectChangeEvent) => {
-        setDepartureStationChoices(event.target.value as string);
-    };
 
     return (
         <div className = "station-wrapper">
             <div className = "dropdown-selections">
                 <div>
-                    <Box sx = { { minWidth: 180 } }>
-                        <FormControl fullWidth>
-                            <InputLabel id = "departure-stations-label" >Departure station</InputLabel>
-
-                            <Select
-                                labelId = "departure-stations-label"
-                                id = "departure-stations"
-                                value = { departureStationChoice }
-                                label = "Departure station"
-                                onChange = { (e: SelectChangeEvent) => (setDepartureStationChoices(e.target.value as string)) }
-                            > 
-                                { stationList.map((stationObj, index) => (
-                                    // Station({ key: index, value: stationObj.value, label: stationObj.label })
-                                    <MenuItem 
-                                        key = { index } 
-                                        value = { stationObj.value }
-                                    >
-                                        { stationObj.label }
-                                    </MenuItem>
-                                )) }
-                            </Select>
-                        </FormControl>
-                    </Box>
-
+                    <StationDropdown 
+                        stationChoice = { departureStationChoice }
+                        setStationChoice = { setDepartureStationChoices }
+                        inputLabel = "Departure station"
+                        id = "departure-stations"  
+                    /> 
                 </div>
                 <div>
-                    <Box sx = { { minWidth: 180 } }>
-                        <FormControl fullWidth>
-                            <InputLabel id = "arrival-stations-label" >Arrival station</InputLabel>
-
-                            <Select
-                                labelId = "arrival-stations-label"
-                                id = "arrival-stations"
-                                value = { arrivalStationChoice }
-                                label = "Arrival station"
-                                onChange = { (e) => (setArrivalStationChoices(e.target.value as string)) }
-                            > 
-                                { stationList.map((stationObj, index) => (
-                                    <MenuItem 
-                                        key = { index } 
-                                        value = { stationObj.value }
-                                    >
-                                        { stationObj.label }
-                                    </MenuItem>
-                                )) }
-                            </Select>
-                        </FormControl>
-                    </Box>
+                    <StationDropdown 
+                        stationChoice = { arrivalStationChoice }
+                        setStationChoice = { setArrivalStationChoices }
+                        inputLabel = "Arrival station"
+                        id = "arrival-stations"                    
+                    />
                 </div>
             </div>
             <button
